Refactor CDR reading into modules
